backend/requirements.txt
# FastAPI and dependencies
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-multipart==0.0.9

# HTTP middleware and CORS
starlette==0.36.3
python-jose[cryptography]==3.3.0

# Environment variables
python-dotenv

# Anthropic API client
anthropic==0.18.1

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2024.1


# Testing dependencies
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0

# Type checking
mypy==1.8.0
types-requests==2.31.0.20240125
types-python-dateutil==2.8.19.14

# Code formatting and linting
black==24.1.1
flake8==7.0.0
isort==5.13.2